+++ winter/core/json/__init__.py
@@ -2,3 +2,4 @@ from .decoder import JSONDecodeException
 from .decoder import json_decode
 from .decoder import json_decoder
 from .encoder import JSONEncoder
+from .undefined import Undefined