+++ winter_django/body_with_context.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from typing import Any
-from typing import Dict
-
-
-@dataclass
-class BodyWithContext:
-    body: Any
-    context: Dict