+++ winter_openapi/inspectors/page_inspector.py
@@ -16,14 +16,15 @@ def inspect_page(hint_class) -> TypeInfo:
     child_class = args[0] if args else str
     extra_fields = set(dataclasses.fields(hint_class.__origin__)) - set(dataclasses.fields(Page))
     child_type_info = inspect_type(child_class)
+    title = child_type_info.title or child_type_info.type_.capitalize()
 
     return TypeInfo(
         type_=DataTypes.OBJECT,
-        title=f'PageOf{child_type_info.title or child_type_info.type_.capitalize()}',
+        title=f'PageOf{title}',
         properties={
             'meta': TypeInfo(
                 type_=DataTypes.OBJECT,
-                title='PageMeta',
+                title=f'PageMetaOf{title}',
                 properties={
                     'total_count': TypeInfo(type_=DataTypes.INTEGER),
                     'limit': TypeInfo(type_=DataTypes.INTEGER, nullable=True),